Started writing directory handler
Added line for botan version in readme just so i won't forget

Started writing a directory class for easy handling of directories.

Can't properly differentiate between files and folders yet.

And the sorting of folder and filenames are case sensetive, might not
want that so will fix after the files and folders problem.